Data Extractions

Functions to extract data from pre-declared filepaths. NEED TO IMPLEMENT CODE TO READ FROM AN INPUT CONFIG FILE.

DataExtractions.extract_data_netcrop — Function

extract_data_netcrop(ISO::String, YEAR_START::Int64, YEAR_END::Int64)

Read from pre-declared directories of CSV files and extract required raw data to perform model analysis on. Extraction code also handles the required upsampling, interpolation and averaging. Output is returned as a Dict. A copy of the extraction is also saved in the listed output directory as a JLD2 file with name format: e.g. AGO\_2010\_2021_extract.jld2

DataExtractions.extract_data_netaccess — Function

extract_data_netaccess(ISO::String, YEAR_START::Int64, YEAR_END::Int64)

Read from pre-declared directories of CSV files and extract required raw data to perform model analysis on. Similar as extract_data() but specifically outputs data needed for calculating net access.

n_max: Maximum number of nets that are considered in aggregated data

h_max: Maximum household size that is considered in aggregated data

DataExtractions.extract_data_netaccess

DataExtractions.extract_data_netcrop