Net Crop Model

Model for estimating national net crop from survey data. Adapted in a reduced form from Bertozzi-Villa, 2021.

NetCropModel.monthly_reduced_itn — Function

monthly_reduced_itn(YEARS_ANNUAL, MONTHS_MONTHLY,
                                DELIVERIES_ANNUAL, DISTRIBUTION_ANNUAL,
                                NET_CROP_MONTHLY; 
                                monthly_p = nothing)

Function of statistical evolution model for net crop component of the stock and flow model. This function uses the @model macro to pass the function into the Sampler construction in Turing.jl. The variable monthly_p contains real values for the relative proportions of each annual distribution quota that is distributed in each given month.

NetCropModel.model_evolve_forward — Function

model_evolve_forward(YEARS_ANNUAL, MONTHS_MONTHLY,
                            DELIVERIES_ANNUAL, DISTRIBUTION_ANNUAL; 
                            monthly_p = nothing, 
                            ϕ_est = 0.5, b_est = 2, k_est = 4)

Forward simulation of the net crop model as above. Requires parameter values for $\phi$, $b$ and $k$ to run. By default, these are set to 0.5, 2 and 4 respectively. If no 'monthly_p' vector is given, a uniform monthly distribution is assumed. Returns a time series as a vector of real numbers of the estimated national net crop.

Net Access Model

NetAccessModel.H_to_access — Function

H_to_access(H::Matrix)

This function takes in a matrix of household counts with $h_{max}$ rows, and $n_{max}-1$ columns. Normalises the entries of $H$ to produce a row stochastic matrix $\hat{H}$ of probabilities on a per individual basis, and subsequently uses it to calulate the net access $\lambda_{\text{access}}$.

NetAccessModel.model_prop_h_nonets — Function

model_prop_h_nonets(emplogit_ρ_h_aggregated::Matrix, γ_aggregated::Matrix)

Turing PPL model for modelling the $\rho_h$, the proportion of households of size $h$ that have no nets available. $emplogit_ρ_h_aggregated$ is a matrix of $h_{max}$ where each row corresponds to a single DHS survey. γ_aggregated is a vector of net per capita values ∈ (0,1) corresponding to the survey of each row in emplogit_ρ_h_aggregated. The regression is modelled with a hierarchical model based on the Bertozzi-Villa and Bhatts model given as follows:

\[\begin{aligned} \beta_0 &\sim \text{Uniform}(-50,50)\\ \beta_1 &\sim \text{Uniform}(-3,3)\\ \beta_2 &\sim \text{Uniform}(-1,1)\\ \beta_3 &\sim \text{Uniform}(-100,100)\\ \beta_4 &\sim \text{Uniform}(-300,300)\\ \beta_5 &\sim \text{Uniform}(-300,300)\\ \tau_\rho &\sim \text{Gamma}(0.1,0.1)\\ \nu_h &= \beta_0 + \beta_1 h + \beta_2 h^2 + \beta_3 \gamma + \beta_4 \gamma^2 + \beta_5 \gamma^3\\ \text{emplogit}(\rho_h) &\sim \text{Normal}(\nu_h, \tau_\rho) \end{aligned}\]

NetAccessModel.model_prop_h_meannets — Function

model_prop_h_meannets(μ_h_aggregated::Matrix, γ_aggregated::Vector)

Turing PPL model for modelling the $\mu_h$, the mean number of net in households of size $h$. Similar to the Bhatt paper, the mean is subsequently used to model entries of the probability matrix $H$ as a zero truncated Poisson distribution. Inputs are matrix of aggregated empirical survey values. μ_h_aggregated is a matrix of $h_{max}$ where each row corresponds to a single DHS survey. γ_aggregated is a vector of net per capita values ∈ (0,1) corresponding to the survey of each row in μ_h_aggregated.

\[\begin{aligned} \beta_0 &\sim \text{Uniform}(-50,50)\\ \beta_1 &\sim \text{Uniform}(-3,3)\\ \beta_2 &\sim \text{Uniform}(-10,10)\\ \beta_3 &\sim \text{Uniform}(-50,50)\\ \beta_4 &\sim \text{Uniform}(-50,50)\\ \beta_5 &\sim \text{Uniform}(-50,50)\\ \tau_\mu &\sim \text{Gamma}(0.1,0.1)\\ \bar{\mu} &= \beta_0 + \beta_1 h + \beta_2 \sqrt{h} + \beta_3 \gamma + \beta_4 \gamma^2 + \beta_5 h \gamma\\ \mu_h &\sim \text{Normal}(\bar{\mu}, \tau_\mu) \end{aligned}\]

NetAccessModel.sample_net_access — Function

sample_net_access(ρ_chain_df::DataFrame, μ_chain_df::DataFrame, p_h::Vector,
                        POPULATION_MONTHLY::Vector, Γ_MONTHLY_samples::Matrix;
                        n_max = 20)

This function is used to generate posterior draws of net access time series conditioned on a set of posterior samples of monthly net crop. The inputs are defined as:

ρ_chain_df: A DataFrame of MCMC chain draws for the parameters to calculate $\rho_h$

μ_chain_df: A DataFrame of MCMC chain draws for the parameters to calculate $\mu_h$

p_h: A vector of probabilities of length $h_{max}$ for the proportion of people with household size $h$

POPULATION_MONTHLY: Monthly estimated time series of population.

Γ_MONTHLY_samples: $N \times m$ matrix of net crop posterior time series draws where $M_{i,j}$ is the net crop of the $i^{th}$ sample time series in month $m$.

n_max: Maximum number of a nets one can expect in a give household. Default limit is 20.

NetAccessModel.H_to_access

NetAccessModel.model_prop_h_meannets

NetAccessModel.model_prop_h_nonets

NetAccessModel.sample_net_access

NetCropModel.model_evolve_forward

NetCropModel.monthly_reduced_itn