Net Attrition

Various loss functions and their inverses for calculating attrition of ITNs in model.

Bertozzi-Villa Model

NetLoss.net_loss_compact — Function

net_loss_compact(t::Float64, τ::Float64; κ = 20, precis_threshold = 1e-6)

Compute the survival rate α based on the compact ITN loss function based on Bertozzi-Villa, 2021.

\[\mathcal{L} = e^{\kappa \left( 1 - \frac{1}{1-\left(\frac{t}{\tau}\right)^2} \right)}\]

Default value of scale parameter is $\kappa = 20$, default rounding precision is 1e-6.

Arguments

t: the time elapsed in years

τ: characteristic time scale of the sigmoidal function

NetLoss.net_life_compact — Function

net_life_compact(α::Float64, τ::Float64; κ = 20, precis_threshold = 1e-6)

Inverse function of net_loss used to compute the α survival lifetime for ITN nets.

Default value of scale parameter is κ = 20, default rounding precision is 1e-6.

Arguments

α: proportion of nets survived

τ: characteristic time scale of the sigmoidal function

Modified Exponential model

NetLoss.net_loss_exp — Function

net_loss_exp(t::Float64, τ_max::Float64; κ_squared = 20, precis_threshold = 1e-6)

Compute the survival rate α based on the a modified loss function based on Bertozzi-Villa

\[\mathcal{L} = e^{\kappa^2 \left( 1 - \frac{1}{1-\left(\frac{t}{\tau_{\text{max}}}\right)^2} \right)}\]

where $\kappa$ is given by the function kappa_coeff and is defined as

\[\kappa = \frac{\log(\alpha)}{ 1 - \frac{1}{ 1 - \left( \frac{ \tau_{\text{min}} }{ \tau_{\text{max}} } \right)^2 } }\]

NetLoss.net_life_exp — Function

net_life_exp(α::Float64, τ_max::Float64; κ_squared = 20, precis_threshold = 1e-6)

Inverse function for the modified exponential model

Weibull model

NetLoss.net_loss_weibull — Function

net_loss_weibull(t::Float64, b::Float64, k::Float64; precis_threshold = 1e-6)

Two parameter Weibull loss model for net attrition given by

\[\mathcal{L} = e^{-\left( \frac{t}{b} \right)^k }\]

NetLoss.net_life_weibull — Function

net_life_weibull(α::Float64, b::Float64, k::Float64)

Inverse function for the Weibull model

NetLoss.net_life_compact

NetLoss.net_life_exp

NetLoss.net_life_weibull

NetLoss.net_loss_compact

NetLoss.net_loss_exp

NetLoss.net_loss_weibull