Plotting Functions

Some some convenience functions to construct plots based on the output .jld2 files from the modelling pipeline.

PlottingFunctions.regression_timeseries_plot — Function

regression_timeseries_plot(input_dict::Dict, regression_dict::Dict, net_access_input_dict::Dict, net_access_chain::Dict, n_samples = 100, kwargs...)

Takes in data extraction and regression results dictionaries from pipeline functions and returns a tuple of figures.

A plot of net crop, net access, and inferred distribution proportion monthly_p with confidence bounds

A plot of net crop stratified by age measured in quarters, with a maximum age of max_age_quarters (default = 16)

Other notable arguments

n_samples: Number of samples to randomly draw from calculate time series. (default = 100)

PlottingFunctions.netlife_posterior_draws — Function

netlife_posterior_draws(regression_dict::Dict; t_vals = 0:0.01:5)

Calculate posterior draws of the survival curve based on the MCMC chain of the net crop model. Returns two outputs

lifecurve_quantiles: 5%, 50% and 95% lifecurve profiles given in a (3, length(t_vals)) matrix.

halflife_quantiles: 5%, 50% and 95% quantiles of the net half life.

PlottingFunctions.lifecurve_plots — Function

lifecurve_plots(ISO::String, lifecurve_quantiles::Matrix, halflife_quantiles::Vector; t_vals = 0:0.01:5, kwargs...)

Takes in the outputs of netlife_posterior draws and returns a plot of the attrition curve.

PlottingFunctions.lifecurve_plots

PlottingFunctions.netlife_posterior_draws

PlottingFunctions.regression_timeseries_plot