Regression functions

Various functions to run the regression analysis and save the outputs.

Net Crop

NetCropRegression.bayes_GD — Function

bayes_GD(input_data_dict::Dict)

Function that takes in the a dictionary of extracted data (see DataExtraction.extract_data) and performs an custom iterative fitting algorithm to identify model parameters monthly_p, $\phi,b,$ and $k$. See default constant hyperparameter values for algorithm in source code. MCMC proposal mechanism is a random walk with manually tuned multivariate normal proposal distribution. Sampler is run using a combination of Turing and AdvancedMH. Returns a dictionary of the final sampled MCMC chains and fits in each iteration.

Note: Output does not trim off the burn-in period and needs to be done in post. But if save_output=true (by default), then a copy of the final chain with burn-in trimmed, and the metadata is saved to the path specified by chainoutputdir. (Refer to source code for path)

Net Access

NetAccessRegression.bayes_access — Function

bayes_access(access_survey_globaldata::Dict; iterations = 10000, burn_in = 2000, chain_output_dir = chain_output_di)

Function to perform MCMC draws of access model parameters ρ_h and μ_h using empirical survey data. Input variable access_survey_globaldata is a Dict with the following key values:

p_h_globaldata: Matrix of aggregated data for distribution of households of size $h$ for each survey.

ρ_h_globaldata: Matrix of aggregated data of proportion of households of size $h$ with no nets for each survey.

μ_h_globaldata: Matrix of aggregated data of mean number of nets in households of size $h$ given at least one net.

Does not return a result, but instead saves the posterior MCMC chain as DataFrames with the filename netaccesschains.jld2 in the directory indicated by chain_output_dir.

NetAccessRegression.bayes_access

NetCropRegression.bayes_GD